Sync struct to official json schema
